+++ app/src/main/java/com/booksy/data/local/UserEntity.kt
@@ -4,11 +4,9 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "user")
-
 data class UserEntity(
-
     @PrimaryKey
-    val id: Int,
+    val id: Long,
     val email: String,
     val name: String,
     val token: String,
